refactor(app): extract session scroll dock controller
